feat: convert to tiles view
