+++ packages/web-app-app-store/src/views/AppList.vue
@@ -1,8 +1,8 @@
 <template>
   <div class="app-list oc-px">
-    <h2 v-text="$gettext('Apps')" />
-    <oc-list>
-      <app-list-item
+    <h2 v-text="$gettext('App Store')" />
+    <oc-list class="app-tiles">
+      <app-tile
         v-for="app in sortedApps"
         :key="`app-${app.repository.name}-${app.id}`"
         :app="app"
@@ -15,12 +15,12 @@
 <script lang="ts">
 import { computed, defineComponent, unref } from 'vue'
 import { useAppsStore } from '../piniaStores'
-import AppListItem from '../components/AppListItem.vue'
+import AppTile from '../components/AppTile.vue'
 import { storeToRefs } from 'pinia'
 
 export default defineComponent({
   name: 'AppList',
-  components: { AppListItem },
+  components: { AppTile },
   setup() {
     const appsStore = useAppsStore()
     const { apps } = storeToRefs(appsStore)
@@ -39,5 +39,11 @@ export default defineComponent({
 <style lang="scss">
 .app-list {
   overflow: auto;
+
+  .app-tiles {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+    gap: 2rem;
+  }
 }
 </style>
